Add 'friend finder' in list of TODOs.
Friend finder is a feature to quickly see if people, with whom
you chat often, or your friends are online or not. This feature
is inspired with the simplicity to communicate with friends on
Google hangout or Facebook chat. You see a person online and you
talk, unlike in the traditional IRC way to browse through channels
and nicks to find a person.